[+] refactor : add desktop job table component and job filters
